requirements.txt
pygame
gym
gym[box2d]
PyYAML
